requirements.txt
fastapi
uvicorn
requests
biopython